hierarchy: fix ordering of menu entries
Previously the code tried to determine a new "file.order" value by
iterating over all files, looking for the highest order. It then
assigned order+100 to a file which didn't have a an order specified.

This turned out to be a flaw of thinking, because the order in which
files are is non-deterministic.