plugins/hierarchy.py: index _parents{} and _childs{} by File objects
Previously the keys for _parents{} and _childs{} where linktitles.
This made problems when several pages had the same linktitle. Having
two pages with the same linktitle however was perfectly legal, as the
linktitle is supposed to be displayed in sidemenus or breadcrumbs,
when the normal page title is too long.